+++ src/components/Header/index.ts
@@ -1,0 +1,3 @@
+export { default as GNBBar } from "@/components/Header/GNBBar/GNBBar.tsx";
+export { default as Logo } from "@/components/Header/Logo/Logo.tsx";
+export { default as UtilMenu } from "@/components/Header/UtilMenu/UtilMenu.tsx";
