+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from "@/components/Header/index.ts";
+export * from "@/components/common/index.ts";
